model/train/requirements.txt
boto3
pillow
opencv_contrib_python